requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
DateTime==5.5
distro==1.9.0
django-environ==0.11.2
fonttools==4.54.1
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
jiter==0.7.0
keyring==25.5.0
kiwisolver==1.4.7
littleutils==0.2.4
matplotlib==3.9.2
more-itertools==10.5.0
numpy==2.1.3
oauthlib==3.2.2
openai==1.54.3
outdated==0.2.2
packaging==24.2
paho-mqtt==2.1.0
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
RapidFuzz==3.10.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
SecretStorage==3.3.3
setuptools==75.3.0
six==1.16.0
sniffio==1.3.1
thefuzz==0.22.1
tqdm==4.67.0
tuke_openlab==0.0.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uuid==1.30
validators==0.34.0
zope.interface==7.1.1
